+++ src/TransformUtil.h
@@ -10,7 +10,7 @@ namespace UtilSimpleECS {
 		 * @param (x,y) pair representing a world space coordinate used by Transforms
 		 * @returns (x,y) pair representing a pixel location in SDL surface coordinates.
 		 */
-		static std::pair<int, int> worldToScreenSpace(int x, int y);
+		static std::pair<int, int> worldToScreenSpace(double x, double y);
 
 		/**
 		 * Converts from a world space coordinate (x, y) to SDL surface coordinate.
@@ -18,6 +18,6 @@ namespace UtilSimpleECS {
 		 * @param (x,y) pair representing a pixel location in SDL surface coordinates.
 		 * @returns (x,y) pair representing a world space coordinate used by Transforms
 		 */
-		static std::pair<int, int> screenToWorldSpace(int x, int y);
+		static std::pair<double, double> screenToWorldSpace(int x, int y);
 	};
 }
